pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dapd-pipeline"
version = "0.1.0"
description = "Domain-Adaptive Progressive Distillation pipeline for small language models"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "DAPD"}]
dependencies = [
  "accelerate>=0.30.0",
  "datasets>=2.19.0",
  "evaluate>=0.4.2",
  "numpy>=1.26.0",
  "peft>=0.11.0",
  "pyyaml>=6.0.1",
  "scikit-learn>=1.4.2",
  "torch>=2.3.0",
  "transformers>=4.42.0",
  "tqdm>=4.66.0",
  "psutil>=5.9.8",
]

[project.optional-dependencies]
qlora = ["bitsandbytes>=0.43.1"]
dev = ["pytest>=8.2.0", "ruff>=0.5.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
